requirements.txt
# requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data
yfinance>=0.2.28
ccxt>=4.0.0  # For crypto, but can be adapted
ib_async>=0.6.0  # Our primary connection to Interactive Brokers

# Backtesting & Analysis
vectorbt>=0.25.0
backtesting>=0.3.3
ta>=0.10.2  # Technical Analysis library

# Web & Monitoring
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.0